meson_options: Autodetect pycamera support
The pycamera was defaulted to be disabled due to a requirement
on a specialised branch of the pybind dependency.

This requirement on a subproject has been removed and we can
now build the pycamera component provided that the distribution
packages pybind directly which is common.

Enable the pycamera build option to be selected automatically
when both Python and pybind are detected on the system.

Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>