libcamera: v4l2_subdevice: Add code member to MediaBusFormatInfo
To prepare for exposing the MediaBusFormatInfo structure as an internal
API, add a code member to the structure to store the media bus code.
This makes MediaBusFormatInfo usable standalone, without having to
externally associate the code related to the info.

The entries in the mediaBusFormatInfo map are becoming too long, so
split them on multiple lines.

While at it, swap the order of the members to match the PixelFormatInfo
class for consistency.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>