libcamera: v4l2_subdevice: Update to the new kernel routing API
The subdev embedded data support series includes a change to the
VIDIOC_SUBDEV_G_ROUTING and VIDIOC_SUBDEV_S_ROUTING ioctls that impacts
the userspace API.

Update to the new API. Backward compatibility can't be easily preserved
as the ioctl structure size has changed. This is not a major issue, as
the routing API isn't enabled in any upstream kernel.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>