ipa: rpi: controller: Add static getHardwareConfig() helper
Add a static helper function Controller::getHardwareConfig() which
overloads the existing Controller::getHardwareConfig() member function,
but takes a string parameter for the hardware target.

Switch the existing Controller::getHardwareConfig() member function to
use the new overload with the target string taken from the tuning file.

This new static helper will be used in a future commit when the
controller has not been initialised, but the target string is available
from the IPA.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>