ipa: rpi: Use static Controller::getHardwareConfig() helper
The IPA knows the target it is running under, so use the new static
Controller::getHardwareConfig() helper with the explicit target string
to get the HW configuration.

This will be needed for a future commit where the controller class will
not yet be initialised, so calling the Controller::getHardwareConfig()
member function will not be possible.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>