pipeline: ipa: rpi: Add setTuning() to the IPA interface
Add a new setTuning() function to the IPA interface used to load the
tuning file into the controller.

The tuning file is now set by the pipeline handler on camera::Start()
instead of the camera enumration stage. This will allow the application
to set an alternate tuning file through a vendor control on camera::Start()
in a future commit.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>