meson: Fix mismatch in controls and properties generated file names
The header for controls and properties are generated from the
control_ids.h.in and property_ids.h.in templates respectively, and the
generated files are named control_ids.h and property_ids.h.

For sources, the templates are named control_ids.cpp.in and
property_ids.cpp.in, but the output files are named controls_ids.cpp and
properties_ids.cpp. This discrepancy causes confusion. Fix it.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>